refactor(agent): collapse the retry template-method split into ModelInvocationNode (A1)
