+++ src/agent/core/README.md
@@ -5,13 +5,13 @@
 ubiquitous language is visible in the directory layout rather than buried in a
 flat folder.
 
-| Module        | Concern                          | Contents                                                                                                                                                                                                                                                                                                                                                                                                           |
-| ------------- | -------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
-| `definition/` | What an agent **is** (configure) | `AgentDataclass` (settings, prompts, `AgentDefinition`, `AgentCategory`), `AgentConfig` (launch/run configuration + payload), `AgentCycleOptions` (typed template-variable vocabulary + two-channel record)                                                                                                                                                                                                        |
-| `state/`      | Run-state snapshots              | `AgentWorkspaceState`, `AgentState` (run/round snapshots + metrics), `executionRequests` (request validation), and `TaskState` + `agentConfigToTaskState` — **not** a live state model: `AgentConfig` is the run-config vocabulary, and `TaskState` is only the frozen wire shape of the CLI NDJSON `setTaskState` event, projected from `AgentConfig` at that one boundary                                        |
-| `usage/`      | Usage value objects              | `ResponseUsage`, `RunUsageAccumulator`                                                                                                                                                                                                                                                                                                                                                                             |
-| `tools/`      | Tool contracts                   | `ToolTypes` (`ITool`, `IToolRegistry`, `MapToolRegistry`)                                                                                                                                                                                                                                                                                                                                                          |
-| `flows/`      | Shared cycle kernel              | Only what both flow families use: `ModelInvocationNode`, `RetryState`, `CommonCycleTypes`, `postCompactionContext`, `BaseFlowServices`, `FlowTransitions`, `CycleServices`, `IToolUseSession`, `toolCallParsing`. The family-specific flows live with their consumers — `ResponseCycleFlow` under `implementations/flows/reflection/`, `ToolUseRoundFlow` + `toolUseRound/` under `implementations/flows/tooluse/` |
+| Module        | Concern                          | Contents                                                                                                                                                                                                                                                                                                                                                                                                                                      |
+| ------------- | -------------------------------- | --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| `definition/` | What an agent **is** (configure) | `AgentDataclass` (settings, prompts, `AgentDefinition`, `AgentCategory`), `AgentConfig` (launch/run configuration + payload), `AgentCycleOptions` (typed template-variable vocabulary + two-channel record)                                                                                                                                                                                                                                   |
+| `state/`      | Run-state snapshots              | `AgentWorkspaceState`, `AgentState` (run/round snapshots + metrics), `executionRequests` (request validation), and `TaskState` + `agentConfigToTaskState` — **not** a live state model: `AgentConfig` is the run-config vocabulary, and `TaskState` is only the frozen wire shape of the CLI NDJSON `setTaskState` event, projected from `AgentConfig` at that one boundary                                                                   |
+| `usage/`      | Usage value objects              | `ResponseUsage`, `RunUsageAccumulator`                                                                                                                                                                                                                                                                                                                                                                                                        |
+| `tools/`      | Tool contracts                   | `ToolTypes` (`ITool`, `IToolRegistry`, `MapToolRegistry`)                                                                                                                                                                                                                                                                                                                                                                                     |
+| `flows/`      | Shared cycle kernel              | Only what both flow families use: `ModelInvocationNode` (the model call and its retry lifecycle), `CommonCycleTypes`, `postCompactionContext`, `BaseFlowServices`, `FlowTransitions`, `CycleServices`, `IToolUseSession`, `toolCallParsing`. The family-specific flows live with their consumers — `ResponseCycleFlow` under `implementations/flows/reflection/`, `ToolUseRoundFlow` + `toolUseRound/` under `implementations/flows/tooluse/` |
 
 ## Dependency direction
 
@@ -29,8 +29,8 @@ depend on neither. Don't introduce imports that point back outward (e.g.
 This diagram covers dependencies _within_ `core`. Flow files may still call a
 canonical host-agnostic collaborator outside `core` directly instead of
 injecting a second reference to the same run-owned service. For example,
-`RetryState.ts` reads the current session through `@agent/runtime/RunContext`
-and calls `includedModelAccess()` for relay-401 token refresh. This
+`ModelInvocationNode.ts` reads the current session through its run scope and
+calls `includedModelAccess()` for relay-401 token refresh. This
 is the same pattern `@agent/modelHandlers` already uses, not a `core`-specific
 exception. None of this pulls in `vscode` or `packages/*`; it's still
 host-agnostic, just not self-contained within `core`'s own module boundaries.
